+++ src/app/models/account.model.ts
@@ -1,5 +1,7 @@
+import { Key } from '@briebug/ngrx-auto-entity';
+
 export class Account {
-  id: number;
+  @Key id: number;
   customer_id: number;
   accountNumber: number;
   name: string;