+++ src/app/state/customer.state.ts
@@ -1,16 +1,7 @@
 import { buildState, IEntityState } from '@briebug/ngrx-auto-entity';
 import { Customer } from 'models/customer.model';
 
-const { initialState, selectors } = buildState(Customer);
-
-export const {
-  selectAll: allCustomers,
-  selectEntities: customerEntities,
-  selectIds: customerIds,
-  selectTotal: totalCustomers,
-  selectIsLoading: customersLoading,
-  selectCurrentEntity: currentCustomer
-} = selectors;
+export const { initialState, facade: CustomerFacadeBase } = buildState(Customer);
 
 export function customerReducer(state = initialState): IEntityState<Customer> {
   return state;